+++ naumov.andrei/A1/rectangle.hpp
@@ -6,16 +6,16 @@
 class Rectangle: public Shape
 {
 public:
-  float getArea() const;
+  double getArea() const;
   rectangle_t getFrameRect() const;
   void move(const point_t & newCentre);
-  void move(float newX, float newY);
+  void move(double newX, double newY);
 
-  void setWidth(float newWidth);
-  void setHeight(float newHeight);
+  void setWidth(double newWidth);
+  void setHeight(double newHeight);
 protected:
-  float width_;
-  float height_;
+  double width_;
+  double height_;
 private:
 };
